completely migrate new benchmarking code out of frame_support
